Refactor SwapPage and related components to use AssetProps; update price fetching logic and improve event handling in BuyComp and SellComp, corrected findPrice hook to action
